requirements.txt
groq
pyyaml
python-dotenv
duckduckgo-search
langchain-community
